+++ src/components/UI/Card.js
@@ -1,4 +1,4 @@
-import "./css/Card.css";
+import "./Card.css";
 
 export default function Card(props){
     const classes = 'card ' + props.className;
